Day 11/recruiter-app/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-dotenv
openai